requirements.txt
Flask
Pillow
gTTS
numpy
opencv-python
fpdf
